支持多种样切换，添加设置项目

+++ Solution/RecentFile/RegistryPersister.cs
@@ -122,6 +122,11 @@ namespace Solution.RecentFile
         }
         public void Clear()
         {
+            RegistryKey k = Registry.CurrentUser.OpenSubKey(RegistryKey);
+            if (k == null)
+            {
+                return;
+            }
             Registry.CurrentUser.DeleteSubKeyTree(RegistryKey);
         }
